code review: remove useless fragment
