--- app/src/main/res/layout/fragment_workflow.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="ca.uqac.alterra.WorkflowFragment"
-    android:background="@color/colorAccent"
-    android:id="@+id/container">
-
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:text="fragment workflow"></TextView>
-
-</FrameLayout>
